AccelerationTrackingModel
but still wrong :(((
