+++ Perfect Cadence/Perfect Cadence/PaceTrackingModel.swift
@@ -6,33 +6,3 @@
 //
 
 import Foundation
-import CoreMotion
-
-// tracks the speed (but seems sus)
-class PaceTrackingModel: ObservableObject {
-    private let motionManager = CMMotionManager()
-    @Published var runningPace: Double = 0.0
-
-    init() {
-        startTrackingPace()
-    }
-
-    func startTrackingPace() {
-        if motionManager.isAccelerometerAvailable {
-            motionManager.accelerometerUpdateInterval = 1.0
-            motionManager.startAccelerometerUpdates(to: .main) { accelerometerData, error in
-                if let accelerometerData = accelerometerData {
-                    self.runningPace = self.calculatePace(acceleration: accelerometerData.acceleration)
-                }
-            }
-        }
-    }
-
-    private func calculatePace(acceleration: CMAcceleration) -> Double {
-        return sqrt(acceleration.x * acceleration.x + acceleration.y * acceleration.y + acceleration.z * acceleration.z)
-    }
-}
-
-
-
-
